pull in documentation and organize
